requirements.txt
asgiref==3.9.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
Faker==37.12.0
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
urllib3==2.5.0
kafka-python==2.3.0
gunicorn==23.0.0
packaging==25.0
